criado functions , e foi repassada para lá a função de nome sidebar criado funcao que limita numero de posts criado funcao que verifica se o método da home é post e, se for, salva o post adiciona if else python para alterar a classe da imgem de curtir caso ela esteja ou nao na postlike do user adicionado java script que seleciona o src da imagem de acordo com a classe ja-curtido. se tiver, ele coloca o coração
