+++ pagina_web/static/css/post_list.css
@@ -1,88 +1,96 @@
+.post {
+  /* ≃? div pai com todos os elementos do post */
+  margin-top: 5px;
+  width: 100%;
+  display: flex;
+  flex-wrap: wrap;
+  height: auto;
 
-.post{/* ≃? div pai com todos os elementos do post */
-    margin-top: 5px; 
-    width: 100%;
-    display: flex;
-    flex-wrap: wrap;
-    height: auto;
-
-    /* ≃° testes */
-    background-color: white;
-    border-top:  0.02px solid #eff3f4;
-    border-bottom:  0.02px solid #eff3f4;
+  /* ≃° testes */
+  background-color: white;
+  border-top: 0.02px solid #eff3f4;
+  border-bottom: 0.02px solid #eff3f4;
 }
 
 /* ≃≃#--------------------------------------------------------------------------------------------------------------------------------------------------------- Div que contém a foto de perfil */
-.post > .post_fgr_perfil{/* ≃?  Div que contém a foto de perfil e espaça os conteúdo do post*/
-    width: 10%;
-    padding: 0 2px;
-    text-align: center;
+.post > .post_fgr_perfil {
+  /* ≃?  Div que contém a foto de perfil e espaça os conteúdo do post*/
+  width: 10%;
+  padding: 0 2px;
+  text-align: center;
 }
-
 
 /* ≃≃≃# foto de perfil */
-.post_fgr_perfil > .post_teste_foto{/* ≃? imagem de perfil */
-    max-height: 3em;
-    margin-top: 4px;
-    border-radius: 25px;
+.post_fgr_perfil > .post_teste_foto {
+  /* ≃? imagem de perfil */
+  max-height: 3em;
+  margin-top: 4px;
+  border-radius: 25px;
 }
-
 
 /* ≃≃#---------------------------------------------------------------------------------------------------------------------------------------------------------- Div quer contém o conteúdo do post */
 
-.post > .post_content{/* ≃? Div que contém o conteúdo do post */
-    width: 90%;
-    padding: 0em 0.5em 0.6em 0.1em;
+.post > .post_content {
+  /* ≃? Div que contém o conteúdo do post */
+  width: 90%;
+  padding: 0em 0.5em 0.6em 0.1em;
 }
 
 /* ≃≃≃# Cabeçalho com nome e user  e tres pontos*/
-.post .post_autoria{/* ≃? Header que contém usuario e nome */
-    display: flex;
-    width: 100%;
-    margin: 0.6em 0 0.2em 0;
-
+.post .post_autoria {
+  /* ≃? Header que contém usuario e nome */
+  display: flex;
+  width: 100%;
+  margin: 0.6em 0 0.2em 0;
 }
 
-.post_autoria > .post_nome_de_usuario{/* ≃? span com Nome de usuário */
-    font-weight: bold;
-    margin-right: 0.3em;
+.post_autoria > .post_nome_de_usuario {
+  /* ≃? span com Nome de usuário */
+  font-weight: bold;
+  margin-right: 0.3em;
 }
 
-.post_autoria > .post_user, .post_autoria > time{/* ≃? ancora com User e <time> com hora */
-    opacity: 0.8;
-    color: gray;
+.post_autoria > .post_user,
+.post_autoria > time {
+  /* ≃? ancora com User e <time> com hora */
+  opacity: 0.8;
+  color: gray;
 }
-.post_autoria .post_tres_pontos{/* ≃?ancora que tem os tres pontinhos à direita*/
-    display: inline-block;
-    width: 1.2em;
-    margin-left: auto;
-    border: none;
-    background-color: white;
-    cursor: pointer;
+.post_autoria .post_tres_pontos {
+  /* ≃?ancora que tem os tres pontinhos à direita*/
+  display: inline-block;
+  width: 1.2em;
+  margin-left: auto;
+  border: none;
+  background-color: white;
+  cursor: pointer;
 }
 
-.post_autoria .post_tres_pontos > img{/* ≃? Imagem dos tres pontos a direita */
-    width: 100%;
+.post_autoria .post_tres_pontos > img {
+  /* ≃? Imagem dos tres pontos a direita */
+  width: 100%;
 }
 
 /* ≃≃≃# Paragráfo com conteúdo */
-.post_content .post_text{
-    word-break: break-all;
+.post_content .post_text {
+  word-break: break-all;
 }
 
 /* ≃≃≃# sub-barra de interação com curtir, retwittar, e etc. */
-.post_icones_de_interacao{/* ≃? aside com os icones de interação */
-    display: flex;
-    width: 80%;
-    justify-content: space-between ;
+.post_icones_de_interacao {
+  /* ≃? aside com os icones de interação */
+  display: flex;
+  width: 80%;
+  justify-content: space-between;
 
-    /* ≃° testes */
-    height: 1.3em;
-    margin: 8px 0 0 0;
+  /* ≃° testes */
+  height: 1.3em;
+  margin: 8px 0 0 0;
 }
 
-.post_icones_de_interacao > img{ /* ≃? icones em si */
-    height: 90%;
+.post_icones_de_interacao > img {
+  /* ≃? icones em si */
+  height: 90%;
 }
 
 /* ≃#  aplicado à vários conteúdos*/
