+++ pagina_web/templates/home/post_list.html
@@ -1,36 +1,69 @@
+{% load static %} {% if numero_de_posts > 0 %} {% for post in Posts %}
+<article class="post">
+  <!-- ≃≃# Div com a foto de perfil -->
+  <div class="post_fgr_perfil">
+    <img
+      src="{% static 'images/posts/perfil_para_teste/perfil.jpg' %}"
+      class="post_teste_foto"
+    />
+  </div>
 
-{% load static %}
-{% if numero_de_posts > 0 %}
-    {% for post in Posts %}
-        <article class="post">
-        
-            <!-- ≃≃# Div com a foto de perfil -->
-            <div class="post_fgr_perfil"> 
-                <img src="{% static 'images/posts/perfil_para_teste/perfil.jpg' %}" class="post_teste_foto">
-            </div>
+  <!-- ≃≃# Div com o conteúdo -->
+  <div class="post_content">
+    <header class="post_autoria">
+      <span class="post_nome_de_usuario">{{post.user.is_active}}</span>
+      <a class="post_user">@{{ post.user }}</a>
+      <time>.&nbsp;&nbsp;. {{post.data }}</time>
+      <button class="post_tres_pontos">
+        <img src="{% static 'images/posts/icones/tres-pontos.svg' %}" />
+      </button>
+    </header>
 
-            <!-- ≃≃# Div com o conteúdo -->
-            <div class="post_content">
-            <header class="post_autoria"><span class="post_nome_de_usuario">{{post.user.is_active}}</span> <a class="post_user">@{{ post.user }}</a> <time>.&nbsp;&nbsp;. {{post.data }}</time> <button class="post_tres_pontos"><img src="{% static 'images/posts/icones/tres-pontos.svg' %}"></button></header>
+    <!-- ≃≃# texto do usuário -->
+    <p class="post_text">{{post.texto}}</p>
 
-            <!-- ≃≃# texto do usuário -->
-            <p class="post_text">{{post.texto}}</p>
+    <!-- ≃≃# Possíveis imagens, se tiver -->
+    <div>e aqui as possíveis imagens, se tiver</div>
 
-            <!-- ≃≃# Possíveis imagens, se tiver -->
-            <div>
-                e aqui as possíveis imagens, se tiver
-            </div>
+    <!-- ≃≃# icones de baixo -->
+    <aside class="post_icones_de_interacao">
+      <img src="{% static 'images/posts/icones/comentar_icone.svg' %}" />
+      <img src="{% static 'images/posts/icones/retwittar_icone.svg' %}" />
+      {% if post in posts_curtidos %}
+      <img
+        class="post-icone-curtir curtido"
+        src="{% static 'images/posts/icones/ja_curtido.svg' %}"
+      />
+      {% else %} {{posts_curtidos.post}}
+      <img
+        class="post-icone-curtir"
+        src="{% static 'images/posts/icones/curtir_icone.svg' %}"
+      />
+      {% endif %}
 
-        <!-- ≃≃# icones de baixo -->
-            <aside class="post_icones_de_interacao">
-                <img src="{% static 'images/posts/icones/comentar_icone.svg' %}">
-                <img src="{% static 'images/posts/icones/retwittar_icone.svg' %}">
-                <img src="{% static 'images/posts/icones/curtir_icone.svg' %}">
-                <img src="{% static 'images/posts/icones/share_icone.svg' %}">
-            </aside>
-            
-            </div>
-            
-        </article>
-    {% endfor %}
-{% endif %}
+      <img src="{% static 'images/posts/icones/share_icone.svg' %}" />
+    </aside>
+  </div>
+</article>
+{% endfor %} {% endif %}
+<script>
+  function muda_cor_de_curtido(elemento) {
+    if (elemento.classList.contains("curtido")) {
+      elemento.src = "{% static 'images/posts/icones/ja_curtido.svg' %}";
+    } else {
+      elemento.src = "{% static 'images/posts/icones/curtir_icone.svg' %}";
+    }
+  }
+
+  lista_de_imagens_curtir = document.querySelectorAll(".post-icone-curtir");
+
+  lista_de_imagens_curtir.forEach(function (value) {
+    value.addEventListener("click", muda_cor_de_curtido(value));
+    console.log(value.eventList);
+  });
+
+  lista_de_imagens_curtir.forEach(function (value) {
+    value.addEventListener("click", muda_cor_de_curtido(value));
+    console.log(value.eventList);
+  });
+</script>
