+++ pagina_web/views.py
@@ -1,7 +1,7 @@
 from multiprocessing import context
 from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
-from pagina_web.models import Posts
+from pagina_web.models import *
 from pagina_web.forms import form_TextoPost
 from django.http import HttpRequest
 from django.contrib.auth.models import User
@@ -10,6 +10,7 @@ from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
 import json
 import requests
+from pagina_web.functions import *
 
 
 # Create your views here.
@@ -23,39 +24,21 @@ def base(request):
 def home(request):
     article = sidebar("https://newsapi.org/v1/articles?country=br&source=bbc-news&sortBy=top&apiKey=f3ddd76e328349ce8967b46f7703dfad")
     
-    
-    if len(Posts.objects.all()) >= 17:
-        n = 17
-    else:
-        n = len(Posts.objects.all())
-        
+    n = limite_posts(Posts.objects.all())
+    posts_curtidos = retorna_lista_de_posts_curtidos(request, banco_PostLike=PostLike)
     context = {
         'numero_de_posts' : n,
-        'Posts': Posts.objects.all().order_by('-date')[:n+1],
+        'Posts': Posts.objects.all().order_by('-date')[:n],
         'form_texto_post': form_TextoPost(),
         'articles': article, 
+        'posts_curtidos' : posts_curtidos,
            
     }
-    if request.method == 'POST':
-        user= User.objects.get(id = request.user.id)
-        Posts.objects.create(user=user,texto = request.POST['text-input'] ).save()
+    verifica_se_eh_post_e_salva(request, banco_user=User, banco_posts=Posts)
     return render(request,'home/home.html',context)
 
 def menubar(request):
     return render(request, 'home/menubar.html', {})
-
-
-def sidebar(url):
-    url = requests.get(url)
-    
-    text = url.text
-    data = json.loads(text)
-    if data['status'] != 'error':
-        article = data['articles']
-    else:
-        article = "............................................................................."  
-    
-    return article
 
 def explorar(request):
     return render(request, 'home/explorar.html', {})
